demos: migrate to Kanama 0.3.0 (version pin, API deltas, re-adopt addons)
- Default kanamaVersion 0.2.2 -> 0.3.0 (gradle/kanama-demo.gradle.kts,
  tps-demo-kanama/build.gradle.kts).
- API deltas from the 0.3.0 wrappers (int32 returns are now Int, not Long):
  - godot-4-3d-third-person-controller GrenadeLauncher: ShapeCast.getCollisionCount()
    compares against 0 (was 0L).
  - Starter-Kit-City-Builder Builder: GridMap.INVALID_CELL_ITEM.toInt() to match
    the int32 get/setCellItem signatures.
- Re-adopt each demo's addons/kanama/kanama.gdextension from the 0.3.0 desktop
  install (desktop + Android entries; ios.* entries are re-added by the iOS
  adopt flow, installIosAddon).

buildAllScripts green against the 0.3.0 kanama source checkout.

+++ Starter-Kit-City-Builder/kotlin-src/Builder.kt
@@ -143,8 +143,8 @@ class Builder(godotObject: MemorySegment) : KanamaScript<Node3D>(godotObject, ::
 	private fun actionDemolish(gridmapPosition: Vector3i) {
 		if (Input.isActionJustPressed("demolish")) {
 			val grid = requireGridMap()
-			if (grid.getCellItem(gridmapPosition) != GridMap.INVALID_CELL_ITEM) {
-				grid.setCellItem(gridmapPosition, GridMap.INVALID_CELL_ITEM)
+			if (grid.getCellItem(gridmapPosition) != GridMap.INVALID_CELL_ITEM.toInt()) {
+				grid.setCellItem(gridmapPosition, GridMap.INVALID_CELL_ITEM.toInt())
 				playAudio("sounds/removal-a.ogg, sounds/removal-b.ogg, sounds/removal-c.ogg, sounds/removal-d.ogg", -20.0)
 			}
 		}
